requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
invoke==2.2.1
logging==0.4.9.6
paramiko==4.0.0
prompt_toolkit==3.0.52
pycparser==2.23
PyNaCl==1.6.1
PyYAML==6.0.3
questionary==2.1.1
tabulate==0.9.0
wcwidth==0.2.14
